split input functions in different files
