+++ app/templates/layers/page.tsx
@@ -4,13 +4,11 @@ import {useCallback, useState} from 'react';
 import {LayerMode} from '../../../src/app/types/layerMode.types';
 import {Player} from '@remotion/player';
 import {Form, Input} from '../../../src/app/forms/input';
-import {
-	useInputChange,
-	useSelectChange,
-} from '../../../src/app/hooks/onInputChange';
+import {useInputChange} from '../../../src/app/hooks/onInputChange';
 import LayerByMode from '../../../src/app/LayerByMode';
 import {SelectInput} from '../../../src/app/forms/selectInput';
 import {ColorInput} from '../../../src/app/forms/colorInput';
+import {useSelectChange} from '../../../src/app/hooks/onSelectChange';
 
 export default function LayersPage() {
 	const [copied, setCopied] = useState(false);
