+++ app/templates/silhouette/page.tsx
@@ -1,15 +1,13 @@
 'use client';
 
 import {Player} from '@remotion/player';
-import {
-	useInputChange,
-	useSelectChange,
-} from '../../../src/app/hooks/onInputChange';
+import {useInputChange} from '../../../src/app/hooks/onInputChange';
 import {Form, Input} from '../../../src/app/forms/input';
 import {Code} from '../../../src/app/Code';
 import {Silhouette} from '../../../remotion/compositions/templates/silhouette/Silhouette';
 import {Side} from '../../../remotion/compositions/templates/silhouette/Silhouette.type';
 import {SelectInput} from '../../../src/app/forms/selectInput';
+import {useSelectChange} from '../../../src/app/hooks/onSelectChange';
 
 export default function SilhouettePage() {
 	const [title, setTitle] = useInputChange<string>('Example of big title ');
